api/requirements.txt
fastapi[standard]
loguru
pydantic[email]
pydantic_settings
alembic
greenlet
asyncpg
sqlalchemy
psycopg2-binary
apscheduler
passlib
PyJWT
aiofiles
